+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Breaking changes:**
 - Removed deprecated API. 
 - Removed `Range` and `Position` from the model, added `TextOp` to list changes and matching parts in the license.
+- Removed `packageName` from `SdkConstraintStatus.fromSdkVersion` constructor.
 
 ## 0.22.24